+++ Makefile
@@ -80,6 +80,7 @@ SRC_GAME			=	win_render_loop.c			\
 						update_medi_ammo.c			\
 						update_enemy_move.c			\
 						update_bullet.c				\
+						update_bullet_2.c			\
 						update_door.c
